# doc/swagger.json
openapi: 3.0.1
info:
  title: LAB-07
  description: Code Fellows lab-07
  version: '0.1'
servers:
  - url: 'http://localhost:8080'
paths:
  /categories/{id}:
    put:
      description: Update an entry at the given id
      parameters:
        - in: path
          name: id   # Note the name is the same as in the path
          required: true
          schema:
            type: integer
            minimum: 1
          description: The entry ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
            example:
                  {"name":"Lisa"}
      responses:
        '200':
          description: The updated entry
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                example:
                    {"id": 1,"name":"Lisa"}
      servers:
        - url: 'http://localhost:8080'
    delete:
      description: Delete an entry at the given id
      parameters:
      - in: path
        name: id   # Note the name is the same as in the path
        required: true
        schema:
          type: integer
          minimum: 1
          description: The entry ID
      responses:
        '200':
          description: Returns the collection with deleted entry removed
          content:
            application/json; charset=utf-8:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                example:
                    []
      servers:
        - url: 'http://localhost:8080'
    servers:
      - url: 'http://localhost:8080'
  /categories:
    get:
      description: Get all the entries
      responses:
        '200':
          description: A collection of all entries
          content:
            application/json; charset=utf-8:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                example:
                    [{"id": 1,"name":"Lisa"}]
      servers:
        - url: 'http://localhost:8080'
    post:
      description: Add a new entry
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
            example:
                  {"name":"Lisa"}
      responses:
        '200':
          description: The newly added entry
          content:
            application/json; charset=utf-8:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                example:
                    {"id": 1,"name":"Lisa"}
      servers:
        - url: 'http://localhost:8080'
    servers:
      - url: 'http://localhost:8080'